+++ lib/src/response/states/apiError/api_error.dart
@@ -4,7 +4,7 @@ import 'package:pop_network/src/response/states/apiError/mapped_api_error.dart';
 class ApiError implements ApiResult {
   ApiError({
     required this.statusCode,
-    required this.error,
+    this.error,
     this.path,
     this.timestamp,
   });
@@ -12,5 +12,5 @@ class ApiError implements ApiResult {
   final int statusCode;
   final String? path;
   final String? timestamp;
-  final MappedApiError error;
+  final MappedApiError? error;
 }